+++ src/search/depth_control.h
@@ -17,14 +17,14 @@ class CDepthControl {
     bool go_deeper(const int current_depth) const;
     bool infinite_depth() const { return (dc_depth_in_plies == INFINITE_DEPTH); };
   public:
-    void set_depth(const int depth);
+    void set_depth(const int64_t depth);
     void set_movetime_ms(const int64_t movetime_ms);
-    void set_nodes(const int nodes);
+    void set_nodes(const int64_t nodes);
   private:
     void clear_all();
     bool enough_time_left_for_one_more_iteration() const;
   private:
-    int dc_depth_in_plies;
+    int64_t dc_depth_in_plies;
     int64_t dc_movetime_ms;
     int64_t dc_nodes;
 };
